Fixed incorrect dragging of floor/ceiling textures with "rotationfloor" / "rotationceiling" UDMF property set in GZDoom Visual Mode. Replaced all doubles with floats. Script Editor: Navigator drop-down now updates automatically. Navigator drop-down now also works for external files. Added Navigator drop-down support for Decorate and Modeldef.

+++ Source/Core/Editing/EditMode.cs
@@ -209,7 +209,7 @@ namespace CodeImp.DoomBuilder.Editing
 		public virtual void OnPresentDisplay() { }
 
 		// Processing events
-		public virtual void OnProcess(double deltatime) { }
+		public virtual void OnProcess(float deltatime) { }
 		
 		// Generic events
 		public virtual void OnReloadResources() { }
